requirements.txt
# It is recommended to install PyTorch manually.
# PyTorch >= 2.0 is recommended, but 1.12 and 1.13 is compatible.
# See instructions at https://pytorch.org/get-started/locally/

click
einops==0.6.1
fairseq==0.12.2
gradio==3.47.1
h5py
librosa<0.10.0
lightning>=2.0.0
matplotlib
mido
MonkeyType==23.3.0
numpy  # ==1.23.5
onnx==1.14.0
onnxsim==0.4.31
praat-parselmouth==0.4.3
PyYAML
scipy
tensorflow==1.15.2
tensorboard
tensorboardX
torchmetrics
tqdm
psola